More test cases; AuthOverride support in App
